Phase 5a: reorganize react tests
