--- rebrand/react/src/tests/smoke.test.tsx
@@ -1,7 +1,0 @@
-import {describe, expect, test} from '@jest/globals';
-
-describe('@_linked/react package scaffold', () => {
-  test('jest + ts-jest + jsdom are configured', () => {
-    expect(true).toBe(true);
-  });
-});
